Changed values in MNIST data to floats between 0.0 and 1.0

+++ src/MNIST.h
@@ -8,7 +8,7 @@
 #include <string>
 #include <vector>
 
-typedef std::vector<uint8_t > Image;
+typedef std::vector<float> Image;
 
 struct MNIST_Dataset {
     std::vector<Image> train_images;
